Extract Number decoding into ./decoding/Number and separate functions into represenative files
