+++ 2-Computer-Systems/problems/1-Protobuf-Varint/src/Bitwise/decoding/Number/is32BitsOrLess.ts
@@ -1,0 +1,7 @@
+import { VarInt } from '../../../varint';
+
+function is32BitsOrLess(varInt: VarInt): boolean {
+  return varInt.length <= 10;
+}
+
+export default is32BitsOrLess;
